arch: arm64: disable SWIOTLB [STREAMEIGHT-7635]
We've observed that on 64 bit platforms 6.6 kernel allocates 64MiB of memory
for software I/O translation lookaside buffer (SWIOTLB) as the following lines
are printed during boot:
software IO TLB: area num 4.
software IO TLB: mapped [mem 0x00000000710d7000-0x00000000750d7000] (64MB)

And /proc/iomem now reports 64MiB more of reserved memory compared to
6.1 kernel.

According to the docs [0], the swiotlb is useful in 3 scenarios:
1) DMA above 32-bit DMA address
2) Confidential Computing
3) "untrusted" DMA transfers

According to the LWN article [1] it was also useful for
> Older devices for PC-class hardware, reflecting the history of that
> architecture, were often limited to 24 bits of address space for DMA
> transfers, meaning that they could only access the lower 16MB of memory.

Since our usecases fall into none of the above, let's disable this feature.

If in the future we encounter one of the above usecases (most likely a device)
which fails to do DMA), we can reenable this with a smaller memory footprint,
by setting swiotlb=0 in the cmdline or by modifying IO_TLB_DEFAULT_SIZE.

[0]: https://docs.kernel.org/core-api/swiotlb.html
[1]: https://lwn.net/Articles/940973/

Signed-off-by: Peter Suti <peter.suti@streamunlimited.com>