misc: add sue-pm-switch driver for simple power control via PM
This driver provides a minimal suspend/resume/shutdown hook for toggling
a regulator specified via the vdd-supply devicetree property. It enables
the regulator at probe time, disables it during suspend, and re-enables
it on resume. On system shutdown or reboot, it also disables the regulator
via the .shutdown() callback.

This allows us to power-on/off an USB phy that does not have it's own
device-tree bindings and where the driver is only loaded when the device
is "plugged-in".

Signed-off-by: Martin Pietryka <martin.pietryka@streamunlimited.com>