net: usb: r8152: correctly configure the status LEDs [STREAMEIGHT-7461]
This configures the green LED0 to blink with link activity and the
amber LED1 to indicate the link status/speed.

The value is hardcoded because there is no nice way to configure this
setting from the device-tree, but also we have observed one commercial
adapter that has this chip where the LEDs are also configured wrongly per
default, so this change is also an improvement there.

Signed-off-by: Martin Pietryka <martin.pietryka@streamunlimited.com>