drivers/ptp: add a virtual ptp hw clock [STREAMEIGHT-7642]
The kernel software timestamping of network packets is based on
CLOCK_REALTIME which means that if the clock changes there will be a
discontinuity in timestamps and likewise, when running a PTP daemon as
a follower, clock adjustments to synchronize the local clock with the
PTP master end up changing the systemwide CLOCK_REALTIME.

We want to avoid both of these issues i.e. CLOCK_REALTIME changes by
the user or NTP must not affect network packet timestamping and having
a PTP daemon running must not highjack CLOCK_REALTIME.

This commit implements a ptp clock backed by software timestamping
which is exposed to userspace as a ptp hw clock (PHC) and runs
independently as a base timestamp plus a delta of CLOCK_MONOTONIC_RAW
timestamps.

Signed-off-by: Rui Matos <rui.matos@streamunlimited.com>