clk: meson: axg-audio: fix MCLK effectively restricted to 512*FS
In 7ef6c9 "limit MST_MCLK_DIV to always divide by 20", MCLK got
effectively restricted to 512*FS, as there is no other suitable divisor
element in the tree above MST_MCLK_DIV. This change adds support for
other common multiples of FS, specifically 128*FS and 256*FS. This is
particularly beneficial for customer projects.

Regarding the original concern about slow recalculation of the clock
tree, I investigated the impact using tracing of clock rate requests
before and after this change. The idea is that adding two more entries
should not have that much of a negative impact. The comparison was done
on aplay -c 2 -r 48000 -f S32_LE after startup. The only additional
requests (with timestamps) are summarized below. This is consistently
reproducible.

54.576561: clk_rate_request_start: aud_mst_c_mclk_sel min 0 max 18446744073709551615, parent hifi_pll (491520020)
54.576562: clk_rate_request_start: hifi_pll min 0 max 18446744073709551615, parent hifi_pll_dco (1966080078)
54.576563: clk_rate_request_start: hifi_pll_dco min 0 max 18446744073709551615, parent xtal (24000000)
54.576565: clk_rate_request_done: hifi_pll_dco min 0 max 18446744073709551615, parent xtal (24000000)
54.576566: clk_rate_request_done: hifi_pll min 0 max 18446744073709551615, parent hifi_pll_dco (1980160156)
54.576567: clk_rate_request_done: aud_mst_c_mclk_sel min 0 max 18446744073709551615, parent hifi_pll (495040039)
54.576568: clk_rate_request_start: aud_mst_c_mclk_sel min 0 max 18446744073709551615, parent hifi_pll (491520020)
54.576569: clk_rate_request_start: hifi_pll min 0 max 18446744073709551615, parent hifi_pll_dco (1966080078)
54.576571: clk_rate_request_start: hifi_pll_dco min 0 max 18446744073709551615, parent xtal (24000000)
54.576572: clk_rate_request_done: hifi_pll_dco min 0 max 18446744073709551615, parent xtal (24000000)
54.576572: clk_rate_request_done: hifi_pll min 0 max 18446744073709551615, parent hifi_pll_dco (1976319824)
54.576574: clk_rate_request_done: aud_mst_c_mclk_sel min 0 max 18446744073709551615, parent hifi_pll (494079956)

We can see that the impact is around 13 microseconds, which does not
basically change the overall timing.

The impact was also tested by trying to reproduce the original issue by
changing the drift compensator value with alsamixer (then you could hear
a glitch if you changed it fast enough, so holding up arrow for example).
Based on testing, this patch does not introduce any such regression.

Signed-off-by: Matej Groma <matej.groma@streamunlimited.com>