S210DTS: add suspend pins for sport3 on the streamer [STREAMEIGHT-6688]
We can disable the output clocks towards the PCM9211 because it has an external crystal
for generating input clocks. On the input side we are slave anyway.

However it is important to note that we can't disable or reset the PCM9211
if we want to detect SPDIF input and wake from NW-standby by it.

Signed-off-by: Peter Suti <peter.suti@streamunlimited.com>