ASoC: codecs: pcm9211: move DIR rate check to `.hw_params` callback [STREAMEIGHT-7673]
PipeWire needs to be able to open the device to probe for possible rates
and formats, so let's not return an error inside `pcm9211_startup()` if no
rate was detected (which is the case when nothing is connected). Instead
move the check to open the device with the correct rate into the
`pcm9211_hw_params()` callback.

Signed-off-by: Martin Pietryka <martin.pietryka@streamunlimited.com>