ASoC: codecs: pcm9211: power down DIT when not playing [STREAMEIGHT-7720]
The S/PDIF output is currently always on by default, wasting power.

This change uses the `trigger()` callback to power the DIT on only when
playback is active and power it off when playback stops.

This approach was chosen for its simplicity, as the existing DAPM
implementation is focused on the input path and would be unnecessarily
complex for this output-only change.

Signed-off-by: Peter Suti <peter.suti@streamunlimited.com>
[use PCM9211_SYS_RESET_TXDIS_MASK instead of _SHIFT]
[remove unnecessary break after return]
Signed-off-by: Radek Dostál <radek.dostal@streamunlimited.com>