ASoC: codecs: pcm9211: do not fail `.hw_params()` on mismatched DIR rate [STREAMEIGHT-7673]
It sometimes can naturally happen that the requested rate is mismatched
from the actual DIR rate. The correct way is still to return an error here,
however PipeWire will then put the node/device into an error state which
makes the node unusable.

To prevent this from happening allow a mismatched DIR rate and print
a warning.

Signed-off-by: Martin Pietryka <martin.pietryka@streamunlimited.com>