ASoC: fsl_spdif: defer snd_ctl_notify() to a bottom half [STREAMEIGHT-7860]
The `snd_ctl_notify()` function re-enables interrupts as a side effect
of its internal locking, which is not safe to call from a hard IRQ
handler (top half).

This was causing the following kernel warning:

[ 192.533424] irq 32 handler spdif_isr+0x0/0x18c enabled interrupts
[ 192.539589] WARNING: CPU: 0 PID: 869 at /kernel/irq/handle.c:161 __handle_irq_event_percpu+0x114/0x13c

Resolve this by deferring the `snd_ctl_notify()` call to the interrupt's
bottom half (threaded IRQ), which runs in a non-atomic context.

There is also no need to report the rate on every CNEW event, only on lock/unlock.

Fixes: 92997da1058af0 ("ASoC: fsl_spdif: support adaptive sample rate [STREAMEIGHT-3986]")

Signed-off-by: Peter Suti <peter.suti@streamunlimited.com>