rtkdrivers: dma: free buffers before allocating new ones [STREAMEIGHT-7988]
This means we can't use this function even if memory was allocated with
GFP_ATOMIC because free always sleeps. This is not a problem anymore as
previous commit 1c74808b57ef ("ASoC: realtek: dma: preallocate dma buffers
in prepare [STREAMEIGHT-7990]") is no longer calling load_dma_period
(which triggers calling of dma_async_tx_descriptor) in an atomic context.

Signed-off-by: Peter Suti <peter.suti@streamunlimited.com>